server: moved config.js to config/keys.js
