+++ server/mongoose_setup.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 
-const { mongoUri } = require("./config");
+const { mongoUri } = require("./config/keys");
 
 mongoose.connect(mongoUri, {
   useNewUrlParser: true,
